refactor(motor): split controller API header
